feat: 글로벌 css 적용 및 테스트
