--- src/App.module.scss
@@ -1,14 +1,0 @@
-#root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
-}
-
-.testButton {
-  color: red;
-}
-
-.clicked {
-  background-color: orange;
-}
