+++ src/pages/test/test.module.scss
@@ -1,0 +1,16 @@
+.container {
+  @include flex-center;
+
+  flex-direction: column;
+  width: 100vw;
+  height: 100vh;
+  gap: 10px;
+}
+
+.testButton {
+  color: red;
+}
+
+.clicked {
+  background-color: orange;
+}
