+++ src/styles/_mixins.scss
@@ -1,0 +1,5 @@
+@mixin flex-center {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
