+++ src/styles/global.scss
@@ -1,0 +1,1 @@
+@import './mixins';
